# logtail: optimize JSON processing