… cmd/tailscaled/depaware.txt
… logtail/logtail_test.go