@raggi 2024-04-08
> In drainPending, use a different maxLen depending on lowMem. In lowMem mode, it is better to perform multiple uploads than it is to construct a large body that OOMs the process.

How often do we generally anticipate this to happen, I ask as it trades radio power time too. I think the answer is not often, but we should do the back of the envelope math