@dsnet 2024-04-08
> How often do we generally anticipate this to happen, I ask as it trades radio power time too. I think the answer is not often, but we should do the back of the envelope math

Not very often. In normal use, I can't imagine producing more than 64KiB logs per upload period.

The main situation where the limit would be reached is if `tailscaled` were offline for a long time and something was generating lots of log lines. I suspect the rush of activity once connectivity is establish is the type of situation where we'd like to be more cautious of memory use.